refactor: move gentx parsing in util package
